requirements/base.txt
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.3
tzdata==2022.2
django-allauth==0.51.0
django-crispy-forms==1.14.0
Pillow==9.2.0
django-braces==1.15.0
python-dateutil==2.8.2
djangorestframework==3.14.0
django-extensions-3.2.1
Python==3.9